add player/enemy death animations with VFX, configure damage logic.
add death animations with dissolve effect to Player and Enemy.
add simple Heal Orb object and drop it on enemy death position.
add DamageCaster to the enemy and configure it.
add hurted state to the player and enemy, configure random chance for it.
add hit impact on the player related to hurted state.
make characters invincible for 2s. after being attacked.
configure player swords dropping when player is dead.